rework workspace parsing, fix abort in simple select
