+++ module/taskrun/runcmd.go
@@ -35,6 +35,8 @@ import (
 	"github.com/swaros/manout"
 )
 
+// DirFind returns the best matching part of depending the arguments, what of the stored paths
+// would be the expected one
 func DirFind(args []string) string {
 	useIndex := -1
 	usePath := "."
@@ -82,4 +84,216 @@ func PrintCnPaths(hints bool) {
 			fmt.Println(manout.MessageCln("\t", "this will be the same as ", manout.BoldTag, "cd ", dirhandle.GetDir(count-1)))
 		}
 	}
+}
+
+// ShowPaths : display all stored paths in the workspace
+func ShowPaths(current string) int {
+
+	configure.PathWorker(func(index int, path string) {
+		if path == current {
+			fmt.Println(manout.MessageCln("\t[", manout.ForeLightYellow, index, manout.CleanTag, "]\t", manout.BoldTag, path))
+		} else {
+			fmt.Println(manout.MessageCln("\t ", manout.ForeLightBlue, index, manout.CleanTag, " \t", path))
+		}
+
+	})
+	return len(configure.UsedConfig.Paths)
+}
+
+func GetAllTargets() ([]string, bool) {
+	plainTargets, found := targetsAsMap()
+	template, _, exists, terr := GetTemplate()
+	if terr != nil {
+		return plainTargets, found
+	}
+	if exists {
+		shareds := detectSharedTargetsAsMap(template)
+		plainTargets = append(plainTargets, shareds...)
+	}
+	return plainTargets, exists && found
+}
+
+func detectSharedTargetsAsMap(current configure.RunConfig) []string {
+	var targets []string
+	SharedFolderExecuter(current, func(_, _ string) {
+		sharedTargets, have := targetsAsMap()
+		if have {
+			targets = append(targets, sharedTargets...)
+		}
+	})
+
+	return targets
+}
+
+func ExistInStrMap(testStr string, check []string) bool {
+	for _, str := range check {
+		if strings.TrimSpace(str) == strings.TrimSpace(testStr) {
+			return true
+		}
+	}
+	return false
+}
+
+func targetsAsMap() ([]string, bool) {
+	var targets []string
+	template, _, exists, terr := GetTemplate()
+	if terr != nil {
+		targets = append(targets, terr.Error())
+		return targets, false
+	}
+	if exists {
+		return templateTargetsAsMap(template)
+	}
+	return targets, false
+}
+
+func templateTargetsAsMap(template configure.RunConfig) ([]string, bool) {
+	var targets []string
+	found := false
+
+	if len(template.Task) > 0 {
+		for _, tasks := range template.Task {
+			if !ExistInStrMap(tasks.ID, targets) && (!tasks.Options.Invisible || showInvTarget) {
+				found = true
+				targets = append(targets, strings.TrimSpace(tasks.ID))
+			}
+		}
+	}
+
+	return targets, found
+}
+
+func printTargets() {
+
+	template, path, exists, terr := GetTemplate()
+	if terr != nil {
+		return
+	}
+	if exists {
+		fmt.Println(manout.MessageCln(manout.ForeDarkGrey, "used taskfile:\t", manout.CleanTag, path))
+		fmt.Println(manout.MessageCln(manout.ForeDarkGrey, "tasks count:  \t", manout.CleanTag, len(template.Task)))
+		if len(template.Task) > 0 {
+			fmt.Println(manout.MessageCln(manout.BoldTag, "existing targets:"))
+			taskList, _ := templateTargetsAsMap(template)
+			for _, tasks := range taskList {
+				fmt.Println("\t", tasks)
+			}
+		} else {
+			fmt.Println(manout.MessageCln("that is what we gor so far:"))
+			fmt.Println()
+		}
+
+		sharedTargets := detectSharedTargetsAsMap(template)
+		if len(sharedTargets) > 0 {
+
+			for _, stasks := range sharedTargets {
+				fmt.Println("\t", stasks, manout.MessageCln(manout.ForeDarkGrey, " shared", manout.CleanTag))
+			}
+
+		}
+	} else {
+		fmt.Println(manout.MessageCln(manout.ForeCyan, "no task-file exists. you can create one by ", manout.CleanTag, " contxt create"))
+	}
+}
+
+func printPaths() {
+	dir, err := dirhandle.Current()
+	if err == nil {
+		fmt.Println(manout.MessageCln(manout.ForeWhite, " current directory: ", manout.BoldTag, dir))
+		fmt.Println(manout.MessageCln(manout.ForeWhite, " current workspace: ", manout.BoldTag, configure.UsedConfig.CurrentSet))
+		notWorkspace := true
+		pathColor := manout.ForeLightBlue
+		if !configure.PathMeightPartOfWs(dir) {
+			pathColor = manout.ForeLightMagenta
+		} else {
+			notWorkspace = false
+		}
+		fmt.Println(" contains paths:")
+		configure.PathWorker(func(index int, path string) {
+			template, _, exists, _ := GetTemplate()
+			add := ""
+			if strings.Contains(dir, path) {
+				add = manout.ResetDim + manout.ForeCyan
+			}
+			if dir == path {
+				add = manout.ResetDim + manout.ForeGreen
+			}
+			if exists {
+				outTasks := ""
+				targets, _ := templateTargetsAsMap(template)
+				for _, tasks := range targets {
+					outTasks = outTasks + " " + tasks
+				}
+
+				fmt.Println(manout.MessageCln("       path: ", manout.Dim, " no ", manout.ForeYellow, index, " ", pathColor, add, path, manout.CleanTag, " targets", "[", manout.ForeYellow, outTasks, manout.CleanTag, "]"))
+			} else {
+				fmt.Println(manout.MessageCln("       path: ", manout.Dim, " no ", manout.ForeYellow, index, " ", pathColor, add, path))
+			}
+		})
+		if notWorkspace {
+			fmt.Println()
+			fmt.Println(manout.MessageCln(manout.BackYellow, manout.ForeBlue, " WARNING ! ", manout.CleanTag, "\tyou are currently in none of the assigned locations."))
+			fmt.Println("\t\tso maybe you are using the wrong workspace")
+		}
+		if !showHints {
+			fmt.Println()
+			fmt.Println(manout.MessageCln(" targets can be executes by ", manout.BoldTag, "contxt run <targetname>", manout.CleanTag, "(for the current directory)"))
+			fmt.Println(manout.MessageCln(" a target can also be executed in all stored paths by ", manout.BoldTag, "contxt run -a <targetname>", manout.CleanTag, " independend from current path"))
+		}
+
+		fmt.Println()
+		if !showHints {
+			fmt.Println(manout.MessageCln(" all workspaces:", " ... change by ", manout.BoldTag, "contxt <workspace>", ""))
+		} else {
+			fmt.Println(manout.MessageCln(" all workspaces:"))
+		}
+		configure.WorkSpaces(func(name string) {
+			if name == configure.UsedConfig.CurrentSet {
+				fmt.Println(manout.MessageCln("\t[ ", manout.BoldTag, name, manout.CleanTag, " ]"))
+			} else {
+				fmt.Println(manout.MessageCln("\t  ", name, "   "))
+			}
+		})
+	}
+}
+
+type pathInfo struct {
+	Path         string   // the stored path
+	Targets      []string // all existing targets
+	Active       bool     // this is the active path
+	IsSubDir     bool     // this path is the active or a subdir of current dir
+	HaveTemplate bool     // in this folder a template exists
+}
+
+type workspace struct {
+	CurrentDir string     // the current directory
+	CurrentWs  string     // the name of the current workspace
+	InWs       bool       // flag if the current path is part of the workspace
+	Paths      []pathInfo // all stored paths
+}
+
+func CollectWorkspaceInfos() (workspace, error) {
+	var ws workspace
+	dir, err := dirhandle.Current()
+	if err == nil {
+		ws.CurrentDir = dir
+		ws.CurrentWs = configure.UsedConfig.CurrentSet
+		ws.InWs = configure.PathMeightPartOfWs(dir)
+
+		configure.PathWorker(func(index int, path string) {
+			var pInfo pathInfo
+			pInfo.Path = path
+			template, _, exists, _ := GetTemplate()
+			pInfo.HaveTemplate = exists
+			pInfo.Active = (dir == path)
+			pInfo.IsSubDir = strings.Contains(dir, path)
+			if exists {
+				pInfo.Targets, _ = templateTargetsAsMap(template)
+			}
+			ws.Paths = append(ws.Paths, pInfo)
+
+		})
+		return ws, nil
+	}
+	return ws, err
 }
