+++ module/taskrun/runner.go
@@ -26,7 +26,6 @@ package taskrun
 import (
 	"fmt"
 	"os"
-	"strings"
 
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -38,7 +37,6 @@ import (
 	"github.com/swaros/manout"
 )
 
-//var log = logrus.New()
 var (
 	log = &logrus.Logger{
 		Out:       os.Stdout,
@@ -709,20 +707,6 @@ func MainExecute() {
 
 }
 
-// ShowPaths : display all stored paths in the workspace
-func ShowPaths(current string) int {
-
-	configure.PathWorker(func(index int, path string) {
-		if path == current {
-			fmt.Println(manout.MessageCln("\t[", manout.ForeLightYellow, index, manout.CleanTag, "]\t", manout.BoldTag, path))
-		} else {
-			fmt.Println(manout.MessageCln("\t ", manout.ForeLightBlue, index, manout.CleanTag, " \t", path))
-		}
-
-	})
-	return len(configure.UsedConfig.Paths)
-}
-
 func CallBackOldWs(oldws string) bool {
 	GetLogger().Info("OLD workspace: ", oldws)
 	// get all paths first
@@ -793,102 +777,6 @@ func doMagicParamOne(param string) bool {
 	return result
 }
 
-func GetAllTargets() ([]string, bool) {
-	plainTargets, found := targetsAsMap()
-	template, _, exists, terr := GetTemplate()
-	if terr != nil {
-		return plainTargets, found
-	}
-	if exists {
-		shareds := detectSharedTargetsAsMap(template)
-		plainTargets = append(plainTargets, shareds...)
-	}
-	return plainTargets, exists && found
-}
-
-func detectSharedTargetsAsMap(current configure.RunConfig) []string {
-	var targets []string
-	SharedFolderExecuter(current, func(_, _ string) {
-		sharedTargets, have := targetsAsMap()
-		if have {
-			targets = append(targets, sharedTargets...)
-		}
-	})
-
-	return targets
-}
-
-func ExistInStrMap(testStr string, check []string) bool {
-	for _, str := range check {
-		if strings.TrimSpace(str) == strings.TrimSpace(testStr) {
-			return true
-		}
-	}
-	return false
-}
-
-func targetsAsMap() ([]string, bool) {
-	var targets []string
-	template, _, exists, terr := GetTemplate()
-	if terr != nil {
-		targets = append(targets, terr.Error())
-		return targets, false
-	}
-	if exists {
-		return templateTargetsAsMap(template)
-	}
-	return targets, false
-}
-
-func templateTargetsAsMap(template configure.RunConfig) ([]string, bool) {
-	var targets []string
-	found := false
-
-	if len(template.Task) > 0 {
-		for _, tasks := range template.Task {
-			if !ExistInStrMap(tasks.ID, targets) && (!tasks.Options.Invisible || showInvTarget) {
-				found = true
-				targets = append(targets, strings.TrimSpace(tasks.ID))
-			}
-		}
-	}
-
-	return targets, found
-}
-
-func printTargets() {
-
-	template, path, exists, terr := GetTemplate()
-	if terr != nil {
-		return
-	}
-	if exists {
-		fmt.Println(manout.MessageCln(manout.ForeDarkGrey, "used taskfile:\t", manout.CleanTag, path))
-		fmt.Println(manout.MessageCln(manout.ForeDarkGrey, "tasks count:  \t", manout.CleanTag, len(template.Task)))
-		if len(template.Task) > 0 {
-			fmt.Println(manout.MessageCln(manout.BoldTag, "existing targets:"))
-			taskList, _ := templateTargetsAsMap(template)
-			for _, tasks := range taskList {
-				fmt.Println("\t", tasks)
-			}
-		} else {
-			fmt.Println(manout.MessageCln("that is what we gor so far:"))
-			fmt.Println()
-		}
-
-		sharedTargets := detectSharedTargetsAsMap(template)
-		if len(sharedTargets) > 0 {
-
-			for _, stasks := range sharedTargets {
-				fmt.Println("\t", stasks, manout.MessageCln(manout.ForeDarkGrey, " shared", manout.CleanTag))
-			}
-
-		}
-	} else {
-		fmt.Println(manout.MessageCln(manout.ForeCyan, "no task-file exists. you can create one by ", manout.CleanTag, " contxt create"))
-	}
-}
-
 func runTargets(_ string, targets string) {
 	RunTargets(targets, true)
 }
@@ -904,65 +792,4 @@ func printOutHeader() {
 func printInfo() {
 	printOutHeader()
 	printPaths()
-}
-
-func printPaths() {
-	dir, err := dirhandle.Current()
-	if err == nil {
-		fmt.Println(manout.MessageCln(manout.ForeWhite, " current directory: ", manout.BoldTag, dir))
-		fmt.Println(manout.MessageCln(manout.ForeWhite, " current workspace: ", manout.BoldTag, configure.UsedConfig.CurrentSet))
-		notWorkspace := true
-		pathColor := manout.ForeLightBlue
-		if !configure.PathMeightPartOfWs(dir) {
-			pathColor = manout.ForeLightMagenta
-		} else {
-			notWorkspace = false
-		}
-		fmt.Println(" contains paths:")
-		configure.PathWorker(func(index int, path string) {
-			template, _, exists, _ := GetTemplate()
-			add := ""
-			if strings.Contains(dir, path) {
-				add = manout.ResetDim + manout.ForeCyan
-			}
-			if dir == path {
-				add = manout.ResetDim + manout.ForeGreen
-			}
-			if exists {
-				outTasks := ""
-				targets, _ := templateTargetsAsMap(template)
-				for _, tasks := range targets {
-					outTasks = outTasks + " " + tasks
-				}
-
-				fmt.Println(manout.MessageCln("       path: ", manout.Dim, " no ", manout.ForeYellow, index, " ", pathColor, add, path, manout.CleanTag, " targets", "[", manout.ForeYellow, outTasks, manout.CleanTag, "]"))
-			} else {
-				fmt.Println(manout.MessageCln("       path: ", manout.Dim, " no ", manout.ForeYellow, index, " ", pathColor, add, path))
-			}
-		})
-		if notWorkspace {
-			fmt.Println()
-			fmt.Println(manout.MessageCln(manout.BackYellow, manout.ForeBlue, " WARNING ! ", manout.CleanTag, "\tyou are currently in none of the assigned locations."))
-			fmt.Println("\t\tso maybe you are using the wrong workspace")
-		}
-		if !showHints {
-			fmt.Println()
-			fmt.Println(manout.MessageCln(" targets can be executes by ", manout.BoldTag, "contxt run <targetname>", manout.CleanTag, "(for the current directory)"))
-			fmt.Println(manout.MessageCln(" a target can also be executed in all stored paths by ", manout.BoldTag, "contxt run -a <targetname>", manout.CleanTag, " independend from current path"))
-		}
-
-		fmt.Println()
-		if !showHints {
-			fmt.Println(manout.MessageCln(" all workspaces:", " ... change by ", manout.BoldTag, "contxt <workspace>", ""))
-		} else {
-			fmt.Println(manout.MessageCln(" all workspaces:"))
-		}
-		configure.WorkSpaces(func(name string) {
-			if name == configure.UsedConfig.CurrentSet {
-				fmt.Println(manout.MessageCln("\t[ ", manout.BoldTag, name, manout.CleanTag, " ]"))
-			} else {
-				fmt.Println(manout.MessageCln("\t  ", name, "   "))
-			}
-		})
-	}
 }
